pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "naoto120424", email = "naotonaka0424@gmail.com"},
]
dependencies = [
    "torch>=2.4.0",
    "numpy>=1.26.4",
    "mlflow>=2.15.1",
    "setuptools>=72.2.0",
    "omegaconf>=2.3.0",
    "tqdm>=4.66.5",
    "einops>=0.8.0",
    "timm>=1.0.8",
    "pytorch-lightning>=2.4.0",
    "hydra-core>=1.3.2",
    "hydra-optuna-sweeper>=1.2.0",
    "dill>=0.3.8",
    "optuna>=2.10.1",
    "SQLAlchemy>=2.0.32",
    "python-dotenv>=1.0.1",
    "seaborn>=0.13.2",
    "ipykernel>=6.29.5",
    "jupyterlab>=4.2.5",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
